Testing AddPetModal
Suppression des tests de AddPetModal du fichier DetailAnnonces.test.js
car il n'appartiennent plus à ce fichier

Ajout des tests dans le fichier AddPetModal.test.js car ce sont les
tests du composant AddPetModal et qu'il ne sont plus dans le fichier
Detailannonces.test.js
